only create persistent storage files/dirs when needed
the additional_args file as well as many persistent storage dirs are re-created on each ilab invocation. Only create them if they do not exist.

resolves #1704

+++ src/instructlab/configuration.py
@@ -539,7 +539,8 @@ def ensure_storage_directories_exist():
     ]
 
     for dirpath in dirs_to_make:
-        os.makedirs(dirpath, exist_ok=True)
+        if not os.path.exists(dirpath):
+            os.makedirs(dirpath, exist_ok=True)
 
     additional_args_and_defaults = {
         "learning_rate": 2e-5,
@@ -557,8 +558,11 @@ def ensure_storage_directories_exist():
     }
 
     # create exper_args file for users to see/edit
-    with open(DEFAULTS.TRAIN_ADDITIONAL_OPTIONS_FILE, "w", encoding="utf-8") as outfile:
-        yaml.dump(additional_args_and_defaults, outfile)
+    if not os.path.isfile(DEFAULTS.TRAIN_ADDITIONAL_OPTIONS_FILE):
+        with open(
+            DEFAULTS.TRAIN_ADDITIONAL_OPTIONS_FILE, "w", encoding="utf-8"
+        ) as outfile:
+            yaml.dump(additional_args_and_defaults, outfile)
 
 
 class Lab:
